requirements-dev.txt
PyYAML
icalendar
flask
pygithub
pep8
flake8
jedi
pytest
pudb
vcrpy
